moved most of the game's logic to helper file
